+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/network/NetworkCimWriter.kt
@@ -8,6 +8,8 @@
 
 package com.zepben.evolve.database.sqlite.cim.network
 
+import com.zepben.evolve.cim.extensions.iec61968.metering.PanDemandResponseFunction
+import com.zepben.evolve.cim.extensions.iec61970.base.wires.BatteryControl
 import com.zepben.evolve.cim.iec61968.assetinfo.*
 import com.zepben.evolve.cim.iec61968.assets.*
 import com.zepben.evolve.cim.iec61968.common.*
@@ -15,6 +17,7 @@ import com.zepben.evolve.cim.iec61968.infiec61968.infassetinfo.CurrentTransforme
 import com.zepben.evolve.cim.iec61968.infiec61968.infassetinfo.PotentialTransformerInfo
 import com.zepben.evolve.cim.iec61968.infiec61968.infassetinfo.RelayInfo
 import com.zepben.evolve.cim.iec61968.metering.EndDevice
+import com.zepben.evolve.cim.iec61968.metering.EndDeviceFunction
 import com.zepben.evolve.cim.iec61968.metering.Meter
 import com.zepben.evolve.cim.iec61968.metering.UsagePoint
 import com.zepben.evolve.cim.iec61968.operations.OperationalRestriction
@@ -38,6 +41,8 @@ import com.zepben.evolve.cim.iec61970.infiec61970.feeder.LvFeeder
 import com.zepben.evolve.cim.iec61970.infiec61970.wires.generation.production.EvChargingUnit
 import com.zepben.evolve.database.sqlite.cim.CimWriter
 import com.zepben.evolve.database.sqlite.cim.tables.associations.*
+import com.zepben.evolve.database.sqlite.cim.tables.extensions.iec61968.metering.TablePanDemandResponseFunctions
+import com.zepben.evolve.database.sqlite.cim.tables.extensions.iec61970.base.wires.TableBatteryControls
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.assetinfo.*
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.assets.*
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.common.*
@@ -45,6 +50,7 @@ import com.zepben.evolve.database.sqlite.cim.tables.iec61968.infiec61968.infasse
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.infiec61968.infassetinfo.TablePotentialTransformerInfo
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.infiec61968.infassetinfo.TableRecloseDelays
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.infiec61968.infassetinfo.TableRelayInfo
+import com.zepben.evolve.database.sqlite.cim.tables.iec61968.metering.TableEndDeviceFunctions
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.metering.TableEndDevices
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.metering.TableMeters
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.metering.TableUsagePoints
@@ -78,6 +84,55 @@ import java.sql.SQLException
 class NetworkCimWriter(
     override val databaseTables: NetworkDatabaseTables
 ) : CimWriter(databaseTables) {
+
+    // ################################
+    // # EXTENSIONS IEC61968 METERING #
+    // ################################
+
+    /**
+     * Save the [PanDemandResponseFunction] fields to [TablePanDemandResponseFunctions].
+     *
+     * @param panDemandResponseFunction The [PanDemandResponseFunction] instance to write to the database.
+     *
+     * @return true if the [PanDemandResponseFunction] was successfully written to the database, otherwise false.
+     * @throws SQLException For any errors encountered writing to the database.
+     */
+    @Throws(SQLException::class)
+    fun save(panDemandResponseFunction: PanDemandResponseFunction): Boolean {
+        val table = databaseTables.getTable<TablePanDemandResponseFunctions>()
+        val insert = databaseTables.getInsert<TablePanDemandResponseFunctions>()
+
+        insert.setNullableString(table.KIND.queryIndex, panDemandResponseFunction.kind.name)
+        insert.setNullableInt(table.APPLIANCE.queryIndex, panDemandResponseFunction.appliance?.toInt())
+
+        return saveEndDeviceFunction(table, insert, panDemandResponseFunction, "pan demand response function")
+    }
+
+    // ##################################
+    // # EXTENSIONS IEC61970 BASE WIRES #
+    // ##################################
+
+    /**
+     * Save the [BatteryControl] fields to [TableBatteryControls].
+     *
+     * @param batteryControl The [BatteryControl] instance to write to the database.
+     *
+     * @return true if the [BatteryControl] was successfully written to the database, otherwise false.
+     * @throws SQLException For any errors encountered writing to the database.
+     */
+    @Throws(SQLException::class)
+    fun save(batteryControl: BatteryControl): Boolean {
+        val table = databaseTables.getTable<TableBatteryControls>()
+        val insert = databaseTables.getInsert<TableBatteryControls>()
+
+        insert.setNullableString(table.BATTERY_UNIT_MRID.queryIndex, batteryControl.batteryUnit?.mRID)
+        insert.setNullableDouble(table.CHARGING_RATE.queryIndex, batteryControl.chargingRate)
+        insert.setNullableDouble(table.DISCHARGING_RATE.queryIndex, batteryControl.dischargingRate)
+        insert.setNullableDouble(table.RESERVE_PERCENT.queryIndex, batteryControl.reservePercent)
+        insert.setNullableString(table.CONTROL_MODE.queryIndex, batteryControl.controlMode.name)
+
+        return saveRegulatingControl(table, insert, batteryControl, "battery control")
+    }
 
     // #######################
     // # IEC61968 Asset Info #
@@ -329,6 +384,11 @@ class NetworkCimWriter(
     @Throws(SQLException::class)
     private fun saveAssetInfo(table: TableAssetInfo, insert: PreparedStatement, assetInfo: AssetInfo, description: String): Boolean {
         return saveIdentifiedObject(table, insert, assetInfo, description)
+    }
+
+    @Throws(SQLException::class)
+    private fun saveAssetFunction(table: TableAssetFunctions, insert: PreparedStatement, assetFunction: AssetFunction, description: String): Boolean {
+        return saveIdentifiedObject(table, insert, assetFunction, description)
     }
 
     @Throws(SQLException::class)
@@ -587,6 +647,19 @@ class NetworkCimWriter(
         endDevice.usagePoints.forEach { status = status and saveAssociation(it, endDevice) }
 
         return status and saveAssetContainer(table, insert, endDevice, description)
+    }
+
+    @Throws(SQLException::class)
+    private fun saveEndDeviceFunction(
+        table: TableEndDeviceFunctions,
+        insert: PreparedStatement,
+        endDeviceFunction: EndDeviceFunction,
+        description: String
+    ): Boolean {
+        insert.setNullableString(table.END_DEVICE_MRID.queryIndex, endDeviceFunction.endDevice?.mRID)
+        insert.setNullableBoolean(table.ENABLED.queryIndex, endDeviceFunction.enabled)
+
+        return saveAssetFunction(table, insert, endDeviceFunction, description)
     }
 
     /**
@@ -2024,6 +2097,8 @@ class NetworkCimWriter(
         insert.setNullableDouble(table.MIN_ALLOWED_TARGET_VALUE.queryIndex, regulatingControl.minAllowedTargetValue)
         insert.setNullableDouble(table.RATED_CURRENT.queryIndex, regulatingControl.ratedCurrent)
         insert.setNullableString(table.TERMINAL_MRID.queryIndex, regulatingControl.terminal?.mRID)
+        insert.setNullableDouble(table.CT_PRIMARY.queryIndex, regulatingControl.ctPrimary)
+        insert.setNullableDouble(table.MIN_TARGET_DEADBAND.queryIndex, regulatingControl.minTargetDeadband)
 
         return savePowerSystemResource(table, insert, regulatingControl, description)
     }
@@ -2076,6 +2151,28 @@ class NetworkCimWriter(
         insert.setNullableDouble(table.SECTIONS.queryIndex, shuntCompensator.sections)
 
         return saveRegulatingCondEq(table, insert, shuntCompensator, description)
+    }
+
+    /**
+     * Save the [StaticVarCompensator] fields to [TableStaticVarCompensators].
+     *
+     * @param staticVarCompensator The [StaticVarCompensator] instance to write to the database.
+     *
+     * @return true if the [StaticVarCompensator] was successfully written to the database, otherwise false.
+     * @throws SQLException For any errors encountered writing to the database.
+     */
+    @Throws(SQLException::class)
+    fun save(staticVarCompensator: StaticVarCompensator): Boolean {
+        val table = databaseTables.getTable<TableStaticVarCompensators>()
+        val insert = databaseTables.getInsert<TableStaticVarCompensators>()
+
+        insert.setNullableDouble(table.CAPACITIVE_RATING.queryIndex, staticVarCompensator.capacitiveRating)
+        insert.setNullableDouble(table.INDUCTIVE_RATING.queryIndex, staticVarCompensator.inductiveRating)
+        insert.setNullableDouble(table.Q.queryIndex, staticVarCompensator.q)
+        insert.setNullableString(table.SVC_CONTROL_MODE.queryIndex, staticVarCompensator.svcControlMode.name)
+        insert.setNullableInt(table.VOLTAGE_SET_POINT.queryIndex, staticVarCompensator.voltageSetPoint)
+
+        return saveRegulatingCondEq(table, insert, staticVarCompensator, "static var compensator")
     }
 
     @Throws(SQLException::class)
@@ -2429,7 +2526,7 @@ class NetworkCimWriter(
         insert.setString(table.USAGE_POINT_MRID.queryIndex, usagePoint.mRID)
         insert.setString(table.END_DEVICE_MRID.queryIndex, endDevice.mRID)
 
-        return insert.tryExecuteSingleUpdate("usage point to end device association association")
+        return insert.tryExecuteSingleUpdate("usage point to end device association")
     }
 
     private fun EquipmentContainer.shouldExportContainerContents(): Boolean = when (this) {