+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/network/NetworkDatabaseTables.kt
@@ -11,6 +11,8 @@ package com.zepben.evolve.database.sqlite.cim.network
 import com.zepben.evolve.database.sqlite.cim.CimDatabaseTables
 import com.zepben.evolve.database.sqlite.cim.tables.SqliteTable
 import com.zepben.evolve.database.sqlite.cim.tables.associations.*
+import com.zepben.evolve.database.sqlite.cim.tables.extensions.iec61968.metering.TablePanDemandResponseFunctions
+import com.zepben.evolve.database.sqlite.cim.tables.extensions.iec61970.base.wires.TableBatteryControls
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.assetinfo.*
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.assets.TableAssetOwners
 import com.zepben.evolve.database.sqlite.cim.tables.iec61968.assets.TablePoles
@@ -60,6 +62,7 @@ class NetworkDatabaseTables : CimDatabaseTables() {
             TableAssetOrganisationRolesAssets(),
             TableAssetOwners(),
             TableBaseVoltages(),
+            TableBatteryControls(),
             TableBatteryUnits(),
             TableBreakers(),
             TableBusbarSections(),
@@ -107,6 +110,7 @@ class NetworkDatabaseTables : CimDatabaseTables() {
             TableOperationalRestrictions(),
             TableOrganisations(),
             TableOverheadWireInfo(),
+            TablePanDemandResponseFunctions(),
             TablePerLengthSequenceImpedances(),
             TablePetersenCoils(),
             TablePhotoVoltaicUnits(),
@@ -142,6 +146,7 @@ class NetworkDatabaseTables : CimDatabaseTables() {
             TableStreetlights(),
             TableSubGeographicalRegions(),
             TableSubstations(),
+            TableStaticVarCompensators(),
             TableSwitchInfo(),
             TableSynchronousMachines(),
             TableSynchronousMachinesReactiveCapabilityCurves(),